+++ src/infra/http/apiClent.ts
@@ -1,0 +1,9 @@
+export async function getJason<T>(url: string): Promise<T> {
+  const response = await fetch(url);
+
+  if (!response.ok) {
+    throw new Error(`Request faild with status ${response.status}`);
+  }
+
+  return response.json() as Promise<T>;
+}