+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Unreleased]
 
+### Added 4/6/26
+
+- Work orders list page
+- Work orders query hook using React Query
+- HTTP client and API integration for listing work orders
+
+## [Unreleased]
+
 ### Added 4/2/26
 
 - Initial React + TypeScript app setup using Vite